requirements.txt
streamlit
pandas
numpy
supabase
openpyxl
python-dotenv
requests
python-dateutil
argon2-cffi
argon2-cffi-bindings
numpy-financial